[x64] Implement conditional usage of xmm16-31
When the host system has AVX512F+AVX512VL, then an additional set of 16
non-volatile SIMD registers are available to be utilized. This modifies
the x64 thunk a bit when saving and loading non-volatile registers.

This seems to work well enough to the point that we're getting xbyak
errors due to the emitter trying to utilize xmm16-31 on instructions
like `vpand` without proper EVEX encoding.